requirements.txt
dotenv
gradio
yfinance
google-generativeai